add aggressive inlining annotations to improve performance lift more floating point type parameters to support both float and double

+++ HigherLogics.Algebra/Float32.cs
@@ -6,86 +6,86 @@ using System.Runtime.CompilerServices;
 namespace HigherLogics.Algebra
 {
     /// <summary>
-    /// A native double.
+    /// A native float.
     /// </summary>
-    public readonly struct Float64 : IFloating<double, Float64>
+    public readonly struct Float32 : IFloating<float, Float32>
     {
-        public readonly double Magnitude;
+        public readonly float Magnitude;
 
-        public Float64(double m) =>
+        public Float32(float m) =>
             this.Magnitude = m;
 
-        public Float64 One => Const(1);
+        public Float32 One => Const(1);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Const(double x) => new Float64(x);
+        public Float32 Const(float x) => new Float32(x);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Const(int x) => new Float64(x);
+        public Float32 Const(int x) => new Float32(x);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public int Sign() => Magnitude < 0 ? -1 : 1;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Negate() => new Float64(-Magnitude);
+        public Float32 Negate() => new Float32(-Magnitude);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Add(Float64 rhs) => Magnitude + rhs.Magnitude;
+        public Float32 Add(Float32 rhs) => Magnitude + rhs.Magnitude;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Add(double rhs) => Magnitude + rhs;
+        public Float32 Add(float rhs) => Magnitude + rhs;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Divide(double rhs) => Magnitude / rhs;
+        public Float32 Divide(float rhs) => Magnitude / rhs;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Divide(Float64 rhs) => Magnitude / rhs.Magnitude;
+        public Float32 Divide(Float32 rhs) => Magnitude / rhs.Magnitude;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Reciprocal() => 1 / Magnitude;
+        public Float32 Reciprocal() => 1 / Magnitude;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Multiply(Float64 rhs) => Magnitude * rhs.Magnitude;
+        public Float32 Multiply(Float32 rhs) => Magnitude * rhs.Magnitude;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Multiply(double rhs) => Magnitude * rhs;
+        public Float32 Multiply(float rhs) => Magnitude * rhs;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Subtract(Float64 rhs) => Magnitude - rhs.Magnitude;
+        public Float32 Subtract(Float32 rhs) => Magnitude - rhs.Magnitude;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Subtract(double rhs) => Magnitude - rhs;
+        public Float32 Subtract(float rhs) => Magnitude - rhs;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 SubtractFrom(double lhs) =>
+        public Float32 SubtractFrom(float lhs) =>
             lhs - Magnitude;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Sin() => Math.Sin(Magnitude);
+        public Float32 Sin() => (float)Math.Sin(Magnitude);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 SinDeg() => Math.Sin(Magnitude * Math.PI / 180);
+        public Float32 SinDeg() => (float)Math.Sin(Magnitude * Math.PI / 180);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Cos() => Math.Cos(Magnitude);
+        public Float32 Cos() => (float)Math.Cos(Magnitude);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 CosDeg() => Math.Cos(Magnitude * Math.PI / 180);
+        public Float32 CosDeg() => (float)Math.Cos(Magnitude * Math.PI / 180);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Pow(int k) => Math.Pow(Magnitude, k);
+        public Float32 Pow(int k) => (float)Math.Pow(Magnitude, k);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Abs() => Math.Abs(Magnitude);
+        public Float32 Abs() => Math.Abs(Magnitude);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Exp() => Math.Exp(Magnitude);
+        public Float32 Exp() => (float)Math.Exp(Magnitude);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Float64 Log() => Math.Log(Magnitude);
+        public Float32 Log() => (float)Math.Log(Magnitude);
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static implicit operator Float64(double x) =>
-            new Float64(x);
+        public static implicit operator Float32(float x) =>
+            new Float32(x);
     }
 }
